feat: LNB 공통 컴포넌트화
